// frontend/src/components/types.ts
// types.ts
export interface Member {
    id: number;
    name: string;
    title: string;
    age: number;
    email: string;
    mobile: string;
    status: string;

}

export interface EditMember {
    id: number;
    name: string;
    title: string;
    age: number;
    email: string;
    mobile: string;
    status: string;
}